783 major overhaul of console launcher.  Combine all into a single project + jar
