+++ .gitignore
@@ -4,6 +4,7 @@
 # Ignore Gradle build output directory
 build
 bin
+dist
 
 .vscode
 .classpath
