+++ src/main/java/org/apereo/openequella/adminconsole/config/ProxySettings.java
