+++ src/main/java/org/apereo/openequella/adminconsole/config/ServerProfile.java
