+++ src/main/java/org/apereo/openequella/adminconsole/util/Pair.java
