+++ src/main/java/org/apereo/openequella/adminconsole/util/Triple.java
