+++ src/main/resources/icons/splash.gif
