+++ src/main/resources/icons/windowicon.gif
