+++ src/test/java/org/apereo/openequella/adminconsole/AppTest.java
